ikke slette greier

+++ applications/program/models.py
@@ -267,8 +267,9 @@ class Signup(models.Model):
         old = Signup.objects.filter(session=self.session, participant=self.participant)
         if old.count() == 1:
             old_signup = old[0]
-            self.pk = old_signup.pk
-            old_signup.delete()
+            if old_signup.pk != self.pk:
+                self.pk = old_signup.pk
+                old_signup.delete()
 
         super().validate_unique(*args, **kwargs)
